pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nyc-housing-acquire"
version = "0.1.0"
description = "Data acquisition layer for NYC housing + neighborhood factors"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "requests>=2.28",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
